Add Day 4 solution
This got a little bit complex, so I had to introduce a BingoCard class
to prevent redundant code (because classes / OOP is a great way of
providing common methods for marking a board, summing up the unmarked
numbers, and checking if a card is a winning card). :sweat_smile:

In addition, this is the first solution that uses the parse_arguments()
method (from the previous commit) to simplify stuff, and oh man, it
really does. :sweat_smile: