[2023] Add PuzzleRunner and related infrastructure code
For the C# solutions that'll be implemented in Advent of Code 2023, the
puzzles for each day will inherit from the `IPuzzle` static interface.
Once registered, it'll be registered with the `PuzzleRunner`. Finally,
when the user starts the console app to run the Advent of Code
solutions, they'll need to pass in the following arguments to the
console app in order:
- File path to the input file
- The day they want to run (integer)
- The problem they want to run (either `Problem1` or `Problem2`)

The command line args will be passed the `PuzzleRunner`, which will
figure out which `IPuzzle` should run.