[2023 - Day 4] Implement solutions for Day 4
This one felt relatively easy. 🤔

Part 1:
Use regular expression to parse out the card ID, winning numbers, and
numbers picked. 👀 Afterwards, use set theory to figure out the number
of matches (i.e., intersection between winning numbers and picked
numbers, as the numbers are all unique; if they weren't, then the
duplicates would've been removed and we would had problems. 😅)

Finally, noticing the fact that the scoring is a power of 2, use math to
figure out the score for the current card, sum it up, and then print our
result! Simply enough.

Part 2:
Just like part 1, we use regular expression to parse out the card ID,
winning numbers, and numbers picked. But now we need some way to keep
track of the number of copies of a particular card—no worries! We can
rely on our trusty dictionary (lookup table) to keep track of that,
using the card ID as the key and the number of copies we have (both
original and the copies we got from winning). :smile:

After parsing and figuring out how the number of matches (win), we:
- Get the number of copies we have for the current card we're working
  with. If we don't have any, then we update our dictionary so that 1
  copy is stored for the current card ID. If the current card exists in
  our dictionary, then we increment that (to include our original copy
  we're working with at the moment) into the dictionary.
- Finally, figure out the next card IDs we won as a result of our
  winnings (e.g., if we're on Card 5 and we got 4 winning numbers, then
  we will get Card 6, Card 7, Card 8, and Card 9, as the problem
  statement mentions).
- Then, for those card IDs we won, we update our dictionary with the new
  number of copies we won (i.e., add the number of copies of the current
  card we're working with, as all copies of the scratchcard yields the
  same number/winnings). :smile:

Once we've went through all the scratchcards in our input file, we sum
up the values portion of the dictionary, and then print the result.

Bam! :sparkles: