[2023 - Day 5] Implement solutions for Day 5
This one really tests your (computer's) patience... or at least, it
makes you consider using mathematics instead of brute forcing a
solution! 😅

Part 1:
Once we've parsed the seed numbers, the next part is to build the
mapping. We can easily do the conversion from a mathematic approach by
deriving the start and end range the mapping rule applies to and
calculate the offset that needs to be added to the input number to get
the proper output number.

Once we do all of that, we can:
1. Take the seed input number.
2. Run it through the mapping (i.e., find the mapping rule where the
   input number is within the applicable range, and then add the offset
   calculated; if no rules were found, return the number as-is).
3. Once we've done all of that, find the smallest location number.

Part 2:
This one can't be brute-forced easily. The naïve approach I was
originally thinking was iterate through all the seeds and run it in the
mapping _until_ someone on reddit made me realize a mathematical
approach can be taken still! 😮

In essence, the second part tells us that the seeds we're reading is
actually a pair: the seed number and the length. With that, go through
each pair and figure out their start and end seed numbers (inclusive).
Then, feed it to the `MapRange()` method of the `Mappings` class, which
will take the input start/end range and then return the corresponding
output, but breaking it up (bucketizing?) as needed to satisfy the
mapping rules. Store the range tuples inside a list, and once we're
done procesisng all of the seed ranges, we move on to the next mapping
and feed that list of output tuples into the next mapping.

Eventually, we'll get the final list of location range mappings--now all
we have to do is find the smallest start range, and that'll lead to our
answer!