[2023 - Day 6] Implement solutions for Day 6
This one is deceptively easily once you realize you can treat this
problem as a quadratic equation. 😅

Part 1:
Each column in the file represents a race. We want to figure out how
many numbers of way we can beat the record distance, given the duration
of the race.

For each race, we use the quadratic formula to figure out what the
minimum and maximum hold time was in order for the toy boat to travel
the record distance. Once we get that, we can add/round up the minimum
hold time and subtract/round down the maximum hold time to beat the
current record.

Once we do that, count the number of ways we can do this by subtracting
minHold from maxHold (and adding one to include maxHold again, as this
is an inclusive range), multiply it to our running product, and
eventually, we'll get our answer!

Part 2:
Oh wait! This part states the spaces between the numbers in the space
should be ignored and that we're actually working with _one large_
number.

Remove the spaces between the number, parse it as one giant duration and
distance, and then use the quadratic formula to figure ou the minimum
and maximum hold time needed _to meet_ the current record. Again, either
add/round up the minimum hold time and subtract/round down the maximum
hold time, subtract the min from the max and add one, and we'll get the
number of ways we can beat the current record.

Simple enough, eh? 🤓