[2023] Add AdventOfCode.Tests project
This project will be used for including unit testing for some auxiliary
classes or methods that are needed for solving a puzzle. :smile:

This makes it easier to test logic and assumptions than having to craft
a puzzle input and running the debugger (via the `--debug` flag) to
ensure that the auxiliary class and/or method are working properly.