[2023 - Day 7] Implement solutions for Day 7
Talk about edge cases... so many to consider that this day has a
dedicated unit test class so that all potential edge cases for the
`Hand` class that was created to solve this day works properly. 😅

Special thanks to /u/LxsterGames on the /r/AdventOfCode subreddit for
creating a better example input file with edge cases (available here:
https://www.reddit.com/r/adventofcode/comments/18cr4xr/2023_day_7_better_example_input_not_a_spoiler/).

For both parts, introduce a `Hand` class that is responsible for
figuring out the hand type. The hand type is determined by counting the
number of occurrence for each card and then sorting the cards by their
count first (so that highest frequency card appears first) and then by
their rank into a list. Then, looking at the first item in the list
(which should be the card with the most occurrence), we look at its
occurrence to determine the hand type. For certain cases where we might
need to disambiguate the types (e.g., if the first card appeared 3
times, then it's possible we have a full-house if the remaining cards
are the same or a three-of-a-kind if the remaining cars are unique), we
look at the number of unique cards in our list of card occurrence.
Finally, to compare who's stronger, the `Hand` class provides an
`IsStrongerThan(Hand)` method for checking that, which under the hood,
calls the `CompareTo(Hand)` method (thanks to the fact we implement
`IComparable<Hand>` in this class, we can use LINQ to sort our list of
hands. 👀)

Now for the specific parts:

Part 1:
Go through each line and parse out the `Hand` of cards and the bid. From
there, use LINQ to sort the cards by how strong they are (weakest first,
per the problem statement), use `Select(((Hand, int Bid), int Rank))` to get the
ranking, multiply the bid with the rank, sum up all our numbers, and
print out our final result!

Part 2:
Same as before, but now we need to include jokers, who can "mask"
themselves as another (high-ranking) card and increase the strength of a
hand. The only caveat: if two hands have the same strength/type, then
when comparing the card in the hands, jokers are classified as the lowest of the
lowest compared to all other cards. 🤔

With that, modify the `Hand` class to support jokers and ensure that
when we figure out the hand's type, we consider _all_ edge cases (e.g.,
5 Jokers or 2 Jokers and 3 different cards--see the unit test for more
example 👀).

Then do the same as before: sort, rank, multiply, and sum, and we'll get
our result!