[2023 - Day 8] Implement solutions for Day 8
The first part is simple but the second one is not... at least, you're
in for a long waiting, if you try to brute force it. 😬

Part 1:
First, read in the directions that we need to go through. Then, create a
dictionary whose key is the node and the value are the left and right
paths for this node.

From there, we look for node AAA in our dictionary, follow the
directions we're supposed to take (given the current step), and then
take that direction (i.e., set the current node we're at to either the
left node if we're supposed to left and vice-versa for right), repeating
until we reach node ZZZ and cycling through our directions if we end up
not finding node ZZZ at the end of our direction list.

Part 2:
The naïve approach I initially took is to look for all starting nodes
that ends with "A" and do something similar in Part 1, i.e., cycle
through the directions for each node until _all_ of them are at the end
node, i.e. ends with "Z", but then that took too long...

Thanks to the helpful hints posted on the /r/AdventOfCode subreddit,
people have pointed out that there's something cyclic about reach from
the start node to the end node, i.e., if you were to continue following
the directions provided, you'll reach the end node _again_ in the same
number of steps. Hmm... sounds like a modulo is involved!

So the more sensible approach I implemented taking for this part is to:
1. Figure out how many steps it takes to go from the start node to the
   end node for each starting node that was found.
2. Figure out which starting node has the "largest" number of steps
   to take to get from **A to **Z, and make of a note of that.
3. Use the number of (2) and keep incrementing a "currentStep" variable
   by that number. While doing that, go through the list of starting
   nodes, get their number of step needed to reach their end node, and
   modulo it with the current step. If all modded numbers are 0, then we
   know our current step is _the_ step where all the starting nodes will
   reach their end node _at the same time_.