[2023 - Day 9] Implement solutions for Day 9
Compared to the tougher past few days, this one felt like a cakewalk.
Implementation of this is more or less exactly how the problem statement
states doing it, heh.

Part 1:
Once we parse the historical value for the current line, use a window
function of size 2 to slide through the array and calculate the
difference. Then for this new (difference) of values, slide through and
calculate its difference. We keep repeating this until the values are
all the same, i.e., there are no difference.

While we're doing this, collect the last values of each line.
Eventually, we sum this up to get our extrapolated value, and then add
it to the running sum, getting our result after processing all
historical values. 😎

Part 2:
This is similar to Part 1, except we're doing the reverse: we're
extrapolating from behind, tryign to figure out the previous value of
the historical values recorded. The same approach is used, except we now
collect the first value of each difference of lines we've calculated,
and then using those collected values, find the difference for the
current line until we get to the original historical values.