[2023 - Day 15] Implement Part 2 solution for Day 15
Use a dictionary to represent each box number and the focal lenses (and
label) that are within the box, and do the proper operations that's
needed based on the provided initialization step. 😎

For a Day 15, Part 1 and Part 2 are very simple compared to the previous
days that I've solved so far. 😅