[2024 - Day 1] Add solutions for Part 1 and Part 2 of Day 1
As expected, this one felt pretty easy. 😁

For Part 1, I solved this by reading the two list of numbers into
memory, sorting it, and then using Python's `zip` function to get the
two numbers we need to see how apart they are (i.e., find its
difference), find said difference and store the (absolute value) result
into a running sum, and return it.

For Part 2, we use Python's _fancy_ `collections.Counter` class to count
the occurrence of all the numbers in the second list, use the first list
to see which number we need to lookup in the second list to get its
occurrence, multiply the resulting occurrence by the number we're
working with, and then store the product in a running sum, which we
return once all numbers have been processed.