[2025] Add `createScanner()` method to the `DaySolution` interface
This can be overriden by classes that implements an Advent of Code
solution (by implementing the `DaySolution` interface) to return a
customized `Scanner` that will take whatever `input` is passed into it
(e.g., standard input or a file).

This will be useful in cases when an Advent of Code solution would like
to use the `Scanner` to extract tokens from the input stream that
doesn't use whitespace (e.g., getting numerical tokens that are
separated by commas instead of spaces or new lines).