[2025 - Day 2] Implement Part 1 of Day 2
Thank you, Internet, for reminding me that logarithms can be used for
figuring out the number of digits that makes up a number (instead of the
naive approach of converting the number into a string and counting its
length).

Also, the sample input was certainly sneaky! Who would've guessed that I
would need to use a `long` instead of an `int` for this solution...

And also! Thank goodness Java has the Scanner class. It certainly makes
parsing number super easy. :sweat_smile: