[2025 - Day 2] Implement Part 2 of Day 2
By using the magic of repeating a string twice, removing the first and
last character, and checking to see if our string we repeated *still*
exists (despite the first and last character being removed), we know
this is string composed of a repeated substring.

Thanks Stack Overflow for telling me about this simplified way of
checking to see if a string is composed of a repeated substring. :smile: