[2025 - Day 3] Implement Part 1 and Part 2 of Day 3
These problems are getting tricky... :sweat_smile:

For Part 1 though, take the naive approach of finding the largest digit.
Then look to the left and right of that digit for the next largest
digit. And then figure out whether the largest digit we first found
should be in the tens place or the ones place so that we get a _large_
joltage. Then sum up our decisions and print out the result!

For Part 2... thanks to a certain visualization that I saw on reddit for
helping figure this out. 😎 If we were to read a number from left to
right, we know the digits to the left are more significant (i.e., causes
"bigger" numbers) than the digits on the right. With that in mind, let's
assume the last twelve "batteries" is the largest joltage. The question
is: can we make something even larger? To do that, look at the first
digit that'll make up the largest joltage, and see if we can find a
number even larger than it to the left. If we do, let that battery be
the "first" digit that makes up our largest joltage, and shrink our
search window so that we only consider batteries to the right of the
largest joltage we found and batteries to the left (and including) the
eleventh-to-last battery. Keep repeating until we can't find anymore
batteries that are larger than the current digit.

Whew, that was a mouthful.