[2025 - Day 4] Implement Part 2 of Day 4
Certainly the easiest puzzle this year... :sweat_smile:

This is Part 1 on repeat, i.e., we will find all accessible rolls,
remove them once, and then try to find the next accessible rolls, until
there are no more rolls that can be removed. Make sure to keep track of
the number of rolls removed and BAM—we get our answer!