[2025 - Day 5] Implement Part 1 and Part 2 of Day 5
Part 1 was very easy: Just check to see if the list of ingredient ID
exists in the list of ranges provided (or not). If it doesn't, then you
know it's spoiled! And if you want to be extra fancy, try to union
overlapped ranges so you don't have as much to traverse. 😎

Part 2 is deceptive though, as it exposed an issue that I didn't count
for in Part 1, when I was trying to implement the "check for overlap,
and if overlapped, union it"--what if your single overlap check didn't
count for another range that could've been overlaped?

For example, if the range overlap was:

1-5
4-6
71-89
2-10

Part 1's algorithm would've accounted for "1-5" and "4-6" to make "1-6",
but it wouldn't have accounted for the "2-10" near the end (so that
union really should've been "1-10" (along with the "71-89" in the
middle).

To handle this, I had to change my approach, i.e.:
1. Once all ranges have been read in, sort it by the starting number in
   ascending order.
2. Then starting from the end of the range list, work backwards to check
   if the current range being processed overlaps with the previous
   range.
   A. If it did, then merge it with the previous range and then delete
     the current range.
   B. If it didn't, continue going backwards.

The (2A) would've ensured that we would keep merging overlapping unions
until we can't anymore, as sorting the ranges by the start number in
ascending order would ensure that all overlapped ranges would be
together.